kloud: add admins for kloudctl and other things
